requirements.txt
fastapi[all]~=0.116.1
sqlmodel~=0.0.24
passlib[bcrypt]~=1.7.4
SQLAlchemy~=2.0.41
pydantic~=2.11.7
aiosmtplib~=4.0.1
python-jose~=3.5.0